+++ personal_mnemonic_medium/data_layer/document_ingesters/base.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 from collections.abc import Sequence
 from pathlib import Path
 
-from personal_mnemonic_medium.note_factories.note import Document
+from personal_mnemonic_medium.document_ingesters.note import Document
 
 
 class DocumentFactory(ABC):
