+++ personal_mnemonic_medium/data_layer/document_ingesters/document.py
