+++ personal_mnemonic_medium/data_layer/document_ingesters/markdown_ingester.py
@@ -5,10 +5,10 @@ from pathlib import Path
 
 from tqdm import tqdm
 
-from personal_mnemonic_medium.note_factories.base import (
+from personal_mnemonic_medium.document_ingesters.base import (
     DocumentFactory,
 )
-from personal_mnemonic_medium.note_factories.note import Document
+from personal_mnemonic_medium.document_ingesters.note import Document
 
 
 class MarkdownNoteFactory(DocumentFactory):
