+++ personal_mnemonic_medium/prompt_extractors/base.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from collections.abc import Sequence
 
-from personal_mnemonic_medium.note_factories.note import Document
+from personal_mnemonic_medium.document_ingesters.note import Document
 from personal_mnemonic_medium.prompt_extractors.prompt import Prompt
