+++ personal_mnemonic_medium/prompt_extractors/cloze_extractor.py
@@ -3,7 +3,7 @@ import re
 from collections.abc import Sequence
 from typing import Any
 
-from personal_mnemonic_medium.note_factories.note import Document
+from personal_mnemonic_medium.document_ingesters.note import Document
 from personal_mnemonic_medium.prompt_extractors.base import (
     PromptExtractor,
 )
