+++ personal_mnemonic_medium/prompt_extractors/prompt.py
@@ -1,4 +1,4 @@
-from personal_mnemonic_medium.note_factories.note import Document
+from personal_mnemonic_medium.document_ingesters.note import Document
 
 
 class Prompt:
