+++ personal_mnemonic_medium/tests/exporters/anki/test_card_converter.py
@@ -2,6 +2,13 @@ from collections.abc import Sequence
 from pathlib import Path
 
 from personal_mnemonic_medium.card_pipeline import CardPipeline
+from personal_mnemonic_medium.document_ingesters.base import (
+    DocumentFactory,
+)
+from personal_mnemonic_medium.document_ingesters.markdown import (
+    MarkdownNoteFactory,
+)
+from personal_mnemonic_medium.document_ingesters.note import Document
 from personal_mnemonic_medium.exporters.anki.card_types.base import (
     AnkiCard,
 )
@@ -12,13 +19,6 @@ from personal_mnemonic_medium.exporters.anki.package_generator import (
     AnkiPackageGenerator,
 )
 from personal_mnemonic_medium.exporters.base import CardExporter
-from personal_mnemonic_medium.note_factories.base import (
-    DocumentFactory,
-)
-from personal_mnemonic_medium.note_factories.markdown import (
-    MarkdownNoteFactory,
-)
-from personal_mnemonic_medium.note_factories.note import Document
 from personal_mnemonic_medium.prompt_extractors.base import (
     PromptExtractor,
 )
