+++ personal_mnemonic_medium/tests/exporters/anki/test_package_generator.py
@@ -1,5 +1,6 @@
 from pathlib import Path
 
+from personal_mnemonic_medium.document_ingesters.note import Document
 from personal_mnemonic_medium.exporters.anki.card_types.base import (
     AnkiCard,
 )
@@ -9,7 +10,6 @@ from personal_mnemonic_medium.exporters.anki.card_types.qa import (
 from personal_mnemonic_medium.exporters.anki.package_generator import (
     AnkiPackageGenerator,
 )
-from personal_mnemonic_medium.note_factories.note import Document
 from personal_mnemonic_medium.prompt_extractors.qa_extractor import (
     QAPrompt,
 )
