+++ personal_mnemonic_medium/tests/prompt_extractors/test_qa_prompt_extractor.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import pytest
 
-from personal_mnemonic_medium.note_factories.note import Document
+from personal_mnemonic_medium.document_ingesters.note import Document
 from personal_mnemonic_medium.prompt_extractors.qa_extractor import (
     QAPromptExtractor,
 )
